commit all changes that didn't get added
