+++ Homework/homework.py
