+++ Loops/while-for-loops.py
@@ -1,0 +1,16 @@
+# We use loops to keep doing the same action until a certain condition is met
+# we also can use a loop to iterate through all of the things within a set
+
+# While loop - While the condition is true do x
+
+count=0
+
+while count < 3:
+    print(f"I'm counting! {count}")
+
+# For loop - for all of the things in this list or dictionary do x action
+
+animalList = ["lions", "tigers", "bears"]
+
+for animal in animalList:
+    print(f"I'm not afraid of {animal}")
