--- new-file.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Our New File</title>
-</head>
-<body>
-    <p> this is new code you didn't have in your local repo</p>
-</body>
-</html>
